+++ src/main/kotlin/me/gurkz/superdupermod/data/OfflineTpData.kt
@@ -1,0 +1,18 @@
+/*
+ * Copyright 2026 Gurkan
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+package me.gurkz.superdupermod.data
+
+import net.minecraft.resources.Identifier
+import net.minecraft.world.phys.Vec3
+
+data class OfflineTpData(
+    val leavePosition: Vec3 = Vec3.ZERO,
+    val cameraVector: Vec3 = Vec3.ZERO,
+    val dimension: Identifier? = null
+)